# ade/old_format.go
package ade

type OldFormat struct {
	Name        string `json:"name"`
	Start       string `json:"start"`
	End         string `json:"end"`
	Rooms       string `json:"rooms"`
	Prof        string `json:"prof"`
	Unite       string `json:"unite"`
	Description string `json:"description"`
}